# apps/api_go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Dev Wiki - Developer Knowledge Sharing Platform",
        "title": "Dev Wiki API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dev Wiki Support",
            "url": "https://github.com/lgdlong/dev-wiki",
            "email": "support@devwiki.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/votes": {
            "get": {
                "description": "Retrieve all votes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Get all votes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_go_internal_domain.VoteResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new vote on an entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Create a new vote",
                "parameters": [
                    {
                        "description": "Create Vote DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_go_internal_domain.CreateVoteDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api_go_internal_domain.VoteResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/votes/change": {
            "post": {
                "description": "Change vote type or remove if same type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Change or toggle vote",
                "parameters": [
                    {
                        "description": "Vote DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_go_internal_domain.CreateVoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_go_internal_domain.VoteResponseDTO"
                        }
                    }
                }
            }
        },
        "/votes/entity/{entityType}/{entityId}": {
            "get": {
                "description": "Retrieve all votes for a specific entity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Get votes by entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity Type (video, tutorial, comment)",
                        "name": "entityType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entityId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_go_internal_domain.VoteResponseDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/votes/entity/{entityType}/{entityId}/count": {
            "get": {
                "description": "Get upvote and downvote counts for a specific entity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Get vote counts for entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity Type (video, tutorial, comment)",
                        "name": "entityType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entityId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/votes/user/{userId}": {
            "get": {
                "description": "Retrieve all votes made by a specific user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Get votes by user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_go_internal_domain.VoteResponseDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/votes/user/{userId}/entity/{entityType}/{entityId}": {
            "get": {
                "description": "Retrieve a user's vote on a specific entity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Get user's vote on entity",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Entity Type (video, tutorial, comment)",
                        "name": "entityType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entityId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_go_internal_domain.VoteResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a user's vote on a specific entity",
                "tags": [
                    "votes"
                ],
                "summary": "Remove user's vote on entity",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Entity Type (video, tutorial, comment)",
                        "name": "entityType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entityId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/votes/{id}": {
            "get": {
                "description": "Retrieve a single vote by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Get a vote by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_go_internal_domain.VoteResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a vote by ID",
                "tags": [
                    "votes"
                ],
                "summary": "Delete a vote",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update an existing vote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Update a vote",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Vote DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_go_internal_domain.UpdateVoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_go_internal_domain.VoteResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api_go_internal_domain.CreateVoteDTO": {
            "type": "object",
            "required": [
                "entityId",
                "entityType",
                "userId",
                "voteType"
            ],
            "properties": {
                "entityId": {
                    "type": "integer"
                },
                "entityType": {
                    "$ref": "#/definitions/api_go_internal_domain.EntityType"
                },
                "userId": {
                    "type": "integer"
                },
                "voteType": {
                    "$ref": "#/definitions/api_go_internal_domain.VoteType"
                }
            }
        },
        "api_go_internal_domain.EntityType": {
            "type": "string",
            "enum": [
                "tutorial",
                "video",
                "product"
            ],
            "x-enum-varnames": [
                "EntityTypeTutorial",
                "EntityTypeVideo",
                "EntityTypeProduct"
            ]
        },
...
        "api_go_internal_domain.UpdateVoteDTO": {
            "type": "object",
            "properties": {
                "entityId": {
                    "type": "integer"
                },
                "entityType": {
                    "$ref": "#/definitions/api_go_internal_domain.EntityType"
                },
                "userId": {
                    "type": "integer"
                },
                "voteType": {
                    "$ref": "#/definitions/api_go_internal_domain.VoteType"
                }
            }
        },
...
        "api_go_internal_domain.VoteResponseDTO": {
            "type": "object",
            "properties": {
                "entityId": {
                    "type": "integer"
                },
                "entityType": {
                    "$ref": "#/definitions/api_go_internal_domain.EntityType"
                },
                "id": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                },
                "voteType": {
                    "$ref": "#/definitions/api_go_internal_domain.VoteType"
                }
            }
        },
        "api_go_internal_domain.VoteType": {
            "type": "string",
            "enum": [
                "up",
                "down"
            ],
            "x-enum-varnames": [
                "VoteTypeUp",
                "VoteTypeDown"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}